+++ src/builder.rs
@@ -27,9 +27,67 @@ use crate::{
 
 const MIN_ACTIONS: usize = 2;
 
+/// An enumeration of rules for Orchard bundle construction.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum BundleType {
+    /// A transactional bundle will be padded if necessary to contain at least 2 actions,
+    /// irrespective of whether any genuine actions are required.
+    Transactional(Flags, Anchor),
+    /// A coinbase bundle is required to have no non-dummy spends. No padding is performed.
+    Coinbase,
+}
+
+impl BundleType {
+    /// Returns the number of logical actions that builder will produce in constructing a bundle
+    /// of this type, given the specified numbers of spends and outputs.
+    ///
+    /// Returns an error if the specified number of spends and outputs is incompatible with
+    /// this bundle type.
+    pub fn num_actions(
+        &self,
+        num_spends: usize,
+        num_outputs: usize,
+    ) -> Result<usize, &'static str> {
+        let num_requested_actions = core::cmp::max(num_spends, num_outputs);
+
+        match self {
+            BundleType::Transactional(flags, _) => {
+                if !flags.spends_enabled() && num_spends > 0 {
+                    Err("Spends are disabled, so num_spends must be zero")
+                } else if !flags.outputs_enabled() && num_outputs > 0 {
+                    Err("Outputs are disabled, so num_outputs must be zero")
+                } else {
+                    Ok(core::cmp::max(num_requested_actions, MIN_ACTIONS))
+                }
+            }
+            BundleType::Coinbase => {
+                if num_spends > 0 {
+                    Err("Coinbase bundles have spends disabled, so num_spends must be zero")
+                } else {
+                    Ok(num_outputs)
+                }
+            }
+        }
+    }
+
+    /// Returns the set of flags and the anchor that will be used for bundle construction.
+    pub fn bundle_config(&self) -> (Flags, Anchor) {
+        match self {
+            BundleType::Transactional(flags, anchor) => (*flags, *anchor),
+            BundleType::Coinbase => (Flags::SPENDS_DISABLED, Anchor::empty_tree()),
+        }
+    }
+}
+
 /// An error type for the kinds of errors that can occur during bundle construction.
 #[derive(Debug)]
 pub enum BuildError {
+    /// Spends are disabled for the provided bundle type.
+    SpendsDisabled,
+    /// Spends are disabled for the provided bundle type.
+    OutputsDisabled,
+    /// The anchor provided to this builder doesn't match the Merkle path used to add a spend.
+    AnchorMismatch,
     /// A bundle could not be built because required signatures were missing.
     MissingSignatures,
     /// An error occurred in the process of producing a proof for a bundle.
@@ -42,6 +100,8 @@ pub enum BuildError {
     /// A signature is valid for more than one input. This should never happen if `alpha`
     /// is sampled correctly, and indicates a critical failure in randomness generation.
     DuplicateSignature,
+    /// The bundle being constructed violated the construction rules for the requested bundle type.
+    BundleTypeNotSatisfiable,
 }
 
 impl Display for BuildError {
@@ -53,11 +113,31 @@ impl Display for BuildError {
             ValueSum(_) => f.write_str("Overflow occurred during value construction"),
             InvalidExternalSignature => f.write_str("External signature was invalid"),
             DuplicateSignature => f.write_str("Signature valid for more than one input"),
+            BundleTypeNotSatisfiable => {
+                f.write_str("Bundle structure did not conform to requested bundle type.")
+            }
+            SpendsDisabled => f.write_str("Spends are not enabled for the requested bundle type."),
+            OutputsDisabled => f.write_str("Spends are not enabled for the requested bundle type."),
+            AnchorMismatch => {
+                f.write_str("All spends must share the anchor requested for the transaction.")
+            }
         }
     }
 }
 
 impl std::error::Error for BuildError {}
+
+impl From<halo2_proofs::plonk::Error> for BuildError {
+    fn from(e: halo2_proofs::plonk::Error) -> Self {
+        BuildError::Proof(e)
+    }
+}
+
+impl From<value::OverflowError> for BuildError {
+    fn from(e: value::OverflowError) -> Self {
+        BuildError::ValueSum(e)
+    }
+}
 
 /// An error type for adding a spend to the builder.
 #[derive(Debug, PartialEq, Eq)]
@@ -94,18 +174,6 @@ impl Display for OutputError {
 }
 
 impl std::error::Error for OutputError {}
-
-impl From<halo2_proofs::plonk::Error> for BuildError {
-    fn from(e: halo2_proofs::plonk::Error) -> Self {
-        BuildError::Proof(e)
-    }
-}
-
-impl From<value::OverflowError> for BuildError {
-    fn from(e: value::OverflowError) -> Self {
-        BuildError::ValueSum(e)
-    }
-}
 
 /// Information about a specific note to be spent in an [`Action`].
 #[derive(Debug)]
@@ -155,31 +223,55 @@ impl SpendInfo {
             merkle_path,
         }
     }
+
+    fn has_matching_anchor(&self, anchor: Anchor) -> bool {
+        if self.note.value() == NoteValue::zero() {
+            true
+        } else {
+            let cm = self.note.commitment();
+            let path_root = self.merkle_path.root(cm.into());
+            path_root == anchor
+        }
+    }
 }
 
-/// Information about a specific recipient to receive funds in an [`Action`].
+/// Information about a specific output to receive funds in an [`Action`].
 #[derive(Debug)]
-struct RecipientInfo {
+pub struct OutputInfo {
     ovk: Option<OutgoingViewingKey>,
     recipient: Address,
     value: NoteValue,
-    memo: Option<[u8; 512]>,
+    memo: [u8; 512],
 }
 
-impl RecipientInfo {
+impl OutputInfo {
+    /// Constructs a new OutputInfo from its constituent parts.
+    pub fn new(
+        ovk: Option<OutgoingViewingKey>,
+        recipient: Address,
+        value: NoteValue,
+        memo: Option<[u8; 512]>,
+    ) -> Self {
+        Self {
+            ovk,
+            recipient,
+            value,
+            memo: memo.unwrap_or_else(|| {
+                let mut memo = [0; 512];
+                memo[0] = 0xf6;
+                memo
+            }),
+        }
+    }
+
     /// Defined in [Zcash Protocol Spec § 4.8.3: Dummy Notes (Orchard)][orcharddummynotes].
     ///
     /// [orcharddummynotes]: https://zips.z.cash/protocol/nu5.pdf#orcharddummynotes
-    fn dummy(rng: &mut impl RngCore) -> Self {
+    pub fn dummy(rng: &mut impl RngCore) -> Self {
         let fvk: FullViewingKey = (&SpendingKey::random(rng)).into();
         let recipient = fvk.address_at(0u32, Scope::External);
 
-        RecipientInfo {
-            ovk: None,
-            recipient,
-            value: NoteValue::zero(),
-            memo: None,
-        }
+        Self::new(None, recipient, NoteValue::zero(), None)
     }
 }
 
@@ -187,12 +279,12 @@ impl RecipientInfo {
 #[derive(Debug)]
 struct ActionInfo {
     spend: SpendInfo,
-    output: RecipientInfo,
+    output: OutputInfo,
     rcv: ValueCommitTrapdoor,
 }
 
 impl ActionInfo {
-    fn new(spend: SpendInfo, output: RecipientInfo, rng: impl RngCore) -> Self {
+    fn new(spend: SpendInfo, output: OutputInfo, rng: impl RngCore) -> Self {
         ActionInfo {
             spend,
             output,
@@ -223,15 +315,7 @@ impl ActionInfo {
         let cm_new = note.commitment();
         let cmx = cm_new.into();
 
-        let encryptor = OrchardNoteEncryption::new(
-            self.output.ovk,
-            note,
-            self.output.memo.unwrap_or_else(|| {
-                let mut memo = [0; 512];
-                memo[0] = 0xf6;
-                memo
-            }),
-        );
+        let encryptor = OrchardNoteEncryption::new(self.output.ovk, note, self.output.memo);
 
         let encrypted_note = TransmittedNoteCiphertext {
             epk_bytes: encryptor.epk().to_bytes().0,
@@ -256,24 +340,27 @@ impl ActionInfo {
     }
 }
 
-/// A builder that constructs a [`Bundle`] from a set of notes to be spent, and recipients
+/// Type alias for an in-progress bundle that has no proofs or signatures.
+///
+/// This is returned by [`Builder::build`].
+pub type UnauthorizedBundle<V> = Bundle<InProgress<Unproven, Unauthorized>, V>;
+
+/// A builder that constructs a [`Bundle`] from a set of notes to be spent, and outputs
 /// to receive funds.
 #[derive(Debug)]
 pub struct Builder {
     spends: Vec<SpendInfo>,
-    recipients: Vec<RecipientInfo>,
-    flags: Flags,
-    anchor: Anchor,
+    outputs: Vec<OutputInfo>,
+    bundle_type: BundleType,
 }
 
 impl Builder {
     /// Constructs a new empty builder for an Orchard bundle.
-    pub fn new(flags: Flags, anchor: Anchor) -> Self {
+    pub fn new(bundle_type: BundleType) -> Self {
         Builder {
             spends: vec![],
-            recipients: vec![],
-            flags,
-            anchor,
+            outputs: vec![],
+            bundle_type,
         }
     }
 
@@ -295,51 +382,38 @@ impl Builder {
         note: Note,
         merkle_path: MerklePath,
     ) -> Result<(), SpendError> {
-        if !self.flags.spends_enabled() {
+        let (flags, anchor) = self.bundle_type.bundle_config();
+        if !flags.spends_enabled() {
             return Err(SpendError::SpendsDisabled);
         }
 
+        let spend = SpendInfo::new(fvk, note, merkle_path).ok_or(SpendError::FvkMismatch)?;
+
         // Consistency check: all anchors must be equal.
-        let cm = note.commitment();
-        let path_root = merkle_path.root(cm.into());
-        if path_root != self.anchor {
+        if !spend.has_matching_anchor(anchor) {
             return Err(SpendError::AnchorMismatch);
         }
 
-        // Check if note is internal or external.
-        let scope = fvk
-            .scope_for_address(&note.recipient())
-            .ok_or(SpendError::FvkMismatch)?;
-
-        self.spends.push(SpendInfo {
-            dummy_sk: None,
-            fvk,
-            scope,
-            note,
-            merkle_path,
-        });
+        self.spends.push(spend);
 
         Ok(())
     }
 
     /// Adds an address which will receive funds in this transaction.
-    pub fn add_recipient(
+    pub fn add_output(
         &mut self,
         ovk: Option<OutgoingViewingKey>,
         recipient: Address,
         value: NoteValue,
         memo: Option<[u8; 512]>,
     ) -> Result<(), OutputError> {
-        if !self.flags.outputs_enabled() {
+        let (flags, _) = self.bundle_type.bundle_config();
+        if !flags.outputs_enabled() {
             return Err(OutputError);
         }
 
-        self.recipients.push(RecipientInfo {
-            ovk,
-            recipient,
-            value,
-            memo,
-        });
+        self.outputs
+            .push(OutputInfo::new(ovk, recipient, value, memo));
 
         Ok(())
     }
@@ -353,7 +427,7 @@ impl Builder {
     /// Returns the action output components that will be produced by the
     /// transaction being constructed
     pub fn outputs(&self) -> &Vec<impl OutputView> {
-        &self.recipients
+        &self.outputs
     }
 
     /// The net value of the bundle to be built. The value of all spends,
@@ -372,89 +446,115 @@ impl Builder {
             .iter()
             .map(|spend| spend.note.value() - NoteValue::zero())
             .chain(
-                self.recipients
+                self.outputs
                     .iter()
-                    .map(|recipient| NoteValue::zero() - recipient.value),
+                    .map(|output| NoteValue::zero() - output.value),
             )
             .fold(Some(ValueSum::zero()), |acc, note_value| acc? + note_value)
             .ok_or(OverflowError)?;
         i64::try_from(value_balance).and_then(|i| V::try_from(i).map_err(|_| value::OverflowError))
     }
 
-    /// Builds a bundle containing the given spent notes and recipients.
+    /// Builds a bundle containing the given spent notes and outputs.
     ///
     /// The returned bundle will have no proof or signatures; these can be applied with
     /// [`Bundle::create_proof`] and [`Bundle::apply_signatures`] respectively.
     pub fn build<V: TryFrom<i64>>(
-        mut self,
-        mut rng: impl RngCore,
-    ) -> Result<Bundle<InProgress<Unproven, Unauthorized>, V>, BuildError> {
-        // Pair up the spends and recipients, extending with dummy values as necessary.
-        let pre_actions: Vec<_> = {
-            let num_spends = self.spends.len();
-            let num_recipients = self.recipients.len();
-            let num_actions = [num_spends, num_recipients, MIN_ACTIONS]
-                .iter()
-                .max()
-                .cloned()
-                .unwrap();
+        self,
+        rng: impl RngCore,
+    ) -> Result<Option<UnauthorizedBundle<V>>, BuildError> {
+        bundle(rng, self.spends, self.outputs, self.bundle_type)
+    }
+}
 
-            self.spends.extend(
-                iter::repeat_with(|| SpendInfo::dummy(&mut rng)).take(num_actions - num_spends),
-            );
-            self.recipients.extend(
-                iter::repeat_with(|| RecipientInfo::dummy(&mut rng))
-                    .take(num_actions - num_recipients),
-            );
+/// Builds a bundle containing the given spent notes and outputs.
+///
+/// The returned bundle will have no proof or signatures; these can be applied with
+/// [`Bundle::create_proof`] and [`Bundle::apply_signatures`] respectively.
+pub fn bundle<V: TryFrom<i64>>(
+    mut rng: impl RngCore,
+    mut spends: Vec<SpendInfo>,
+    mut outputs: Vec<OutputInfo>,
+    bundle_type: BundleType,
+) -> Result<Option<UnauthorizedBundle<V>>, BuildError> {
+    let (flags, anchor) = bundle_type.bundle_config();
 
-            // Shuffle the spends and recipients, so that learning the position of a
-            // specific spent note or output note doesn't reveal anything on its own about
-            // the meaning of that note in the transaction context.
-            self.spends.shuffle(&mut rng);
-            self.recipients.shuffle(&mut rng);
+    let num_requested_spends = spends.len();
+    if !flags.spends_enabled() && num_requested_spends > 0 {
+        return Err(BuildError::SpendsDisabled);
+    }
 
-            self.spends
-                .into_iter()
-                .zip(self.recipients.into_iter())
-                .map(|(spend, recipient)| ActionInfo::new(spend, recipient, &mut rng))
-                .collect()
-        };
+    for spend in &spends {
+        if !spend.has_matching_anchor(anchor) {
+            return Err(BuildError::AnchorMismatch);
+        }
+    }
 
-        // Move some things out of self that we will need.
-        let flags = self.flags;
-        let anchor = self.anchor;
+    let num_requested_outputs = outputs.len();
+    if !flags.outputs_enabled() && num_requested_outputs > 0 {
+        return Err(BuildError::OutputsDisabled);
+    }
 
-        // Determine the value balance for this bundle, ensuring it is valid.
-        let value_balance = pre_actions
-            .iter()
-            .fold(Some(ValueSum::zero()), |acc, action| {
-                acc? + action.value_sum()
-            })
-            .ok_or(OverflowError)?;
+    let num_actions = bundle_type
+        .num_actions(num_requested_spends, num_requested_outputs)
+        .map_err(|_| BuildError::BundleTypeNotSatisfiable)?;
 
-        let result_value_balance: V = i64::try_from(value_balance)
-            .map_err(BuildError::ValueSum)
-            .and_then(|i| V::try_from(i).map_err(|_| BuildError::ValueSum(value::OverflowError)))?;
+    // Pair up the spends and outputs, extending with dummy values as necessary.
+    let pre_actions: Vec<_> = {
+        spends.extend(
+            iter::repeat_with(|| SpendInfo::dummy(&mut rng))
+                .take(num_actions - num_requested_spends),
+        );
+        outputs.extend(
+            iter::repeat_with(|| OutputInfo::dummy(&mut rng))
+                .take(num_actions - num_requested_outputs),
+        );
 
-        // Compute the transaction binding signing key.
-        let bsk = pre_actions
-            .iter()
-            .map(|a| &a.rcv)
-            .sum::<ValueCommitTrapdoor>()
-            .into_bsk();
+        // Shuffle the spends and outputs, so that learning the position of a
+        // specific spent note or output note doesn't reveal anything on its own about
+        // the meaning of that note in the transaction context.
+        spends.shuffle(&mut rng);
+        outputs.shuffle(&mut rng);
 
-        // Create the actions.
-        let (actions, circuits): (Vec<_>, Vec<_>) =
-            pre_actions.into_iter().map(|a| a.build(&mut rng)).unzip();
+        spends
+            .into_iter()
+            .zip(outputs.into_iter())
+            .map(|(spend, output)| ActionInfo::new(spend, output, &mut rng))
+            .collect()
+    };
 
-        // Verify that bsk and bvk are consistent.
-        let bvk = (actions.iter().map(|a| a.cv_net()).sum::<ValueCommitment>()
-            - ValueCommitment::derive(value_balance, ValueCommitTrapdoor::zero()))
-        .into_bvk();
-        assert_eq!(redpallas::VerificationKey::from(&bsk), bvk);
+    // Determine the value balance for this bundle, ensuring it is valid.
+    let value_balance = pre_actions
+        .iter()
+        .fold(Some(ValueSum::zero()), |acc, action| {
+            acc? + action.value_sum()
+        })
+        .ok_or(OverflowError)?;
 
-        Ok(Bundle::from_parts(
-            NonEmpty::from_vec(actions).unwrap(),
+    let result_value_balance: V = i64::try_from(value_balance)
+        .map_err(BuildError::ValueSum)
+        .and_then(|i| V::try_from(i).map_err(|_| BuildError::ValueSum(value::OverflowError)))?;
+
+    // Compute the transaction binding signing key.
+    let bsk = pre_actions
+        .iter()
+        .map(|a| &a.rcv)
+        .sum::<ValueCommitTrapdoor>()
+        .into_bsk();
+
+    // Create the actions.
+    let (actions, circuits): (Vec<_>, Vec<_>) =
+        pre_actions.into_iter().map(|a| a.build(&mut rng)).unzip();
+
+    // Verify that bsk and bvk are consistent.
+    let bvk = (actions.iter().map(|a| a.cv_net()).sum::<ValueCommitment>()
+        - ValueCommitment::derive(value_balance, ValueCommitTrapdoor::zero()))
+    .into_bvk();
+    assert_eq!(redpallas::VerificationKey::from(&bsk), bvk);
+
+    Ok(NonEmpty::from_vec(actions).map(|actions| {
+        Bundle::from_parts(
+            actions,
             flags,
             result_value_balance,
             anchor,
@@ -462,8 +562,8 @@ impl Builder {
                 proof: Unproven { circuits },
                 sigs: Unauthorized { bsk },
             },
-        ))
-    }
+        )
+    }))
 }
 
 /// Marker trait representing bundle signatures in the process of being created.
@@ -749,7 +849,7 @@ pub trait OutputView {
     fn value<V: From<u64>>(&self) -> V;
 }
 
-impl OutputView for RecipientInfo {
+impl OutputView for OutputInfo {
     fn value<V: From<u64>>(&self) -> V {
         V::from(self.value.inner())
     }
@@ -777,7 +877,7 @@ pub mod testing {
         Address, Note,
     };
 
-    use super::Builder;
+    use super::{Builder, BundleType};
 
     /// An intermediate type used for construction of arbitrary
     /// bundle values. This type is required because of a limitation
@@ -793,7 +893,7 @@ pub mod testing {
         sk: SpendingKey,
         anchor: Anchor,
         notes: Vec<(Note, MerklePath)>,
-        recipient_amounts: Vec<(Address, NoteValue)>,
+        output_amounts: Vec<(Address, NoteValue)>,
     }
 
     impl<R: RngCore + CryptoRng> ArbitraryBundleInputs<R> {
@@ -801,24 +901,25 @@ pub mod testing {
         fn into_bundle<V: TryFrom<i64>>(mut self) -> Bundle<Authorized, V> {
             let fvk = FullViewingKey::from(&self.sk);
             let flags = Flags::from_parts(true, true);
-            let mut builder = Builder::new(flags, self.anchor);
+            let mut builder = Builder::new(BundleType::Transactional(flags, self.anchor));
 
             for (note, path) in self.notes.into_iter() {
                 builder.add_spend(fvk.clone(), note, path).unwrap();
             }
 
-            for (addr, value) in self.recipient_amounts.into_iter() {
+            for (addr, value) in self.output_amounts.into_iter() {
                 let scope = fvk.scope_for_address(&addr).unwrap();
                 let ovk = fvk.to_ovk(scope);
 
                 builder
-                    .add_recipient(Some(ovk.clone()), addr, value, None)
+                    .add_output(Some(ovk.clone()), addr, value, None)
                     .unwrap();
             }
 
             let pk = ProvingKey::build();
             builder
                 .build(&mut self.rng)
+                .unwrap()
                 .unwrap()
                 .create_proof(&pk, &mut self.rng)
                 .unwrap()
@@ -834,7 +935,7 @@ pub mod testing {
         fn arb_bundle_inputs(sk: SpendingKey)
         (
             n_notes in 1usize..30,
-            n_recipients in 1..30,
+            n_outputs in 1..30,
 
         )
         (
@@ -843,12 +944,12 @@ pub mod testing {
                 arb_positive_note_value(MAX_NOTE_VALUE / n_notes as u64).prop_flat_map(arb_note),
                 n_notes
             ),
-            recipient_amounts in vec(
+            output_amounts in vec(
                 arb_address().prop_flat_map(move |a| {
-                    arb_positive_note_value(MAX_NOTE_VALUE / n_recipients as u64)
+                    arb_positive_note_value(MAX_NOTE_VALUE / n_outputs as u64)
                         .prop_map(move |v| (a, v))
                 }),
-                n_recipients as usize
+                n_outputs as usize
             ),
             rng_seed in prop::array::uniform32(prop::num::u8::ANY)
         ) -> ArbitraryBundleInputs<StdRng> {
@@ -873,7 +974,7 @@ pub mod testing {
                 sk,
                 anchor: frontier.root().into(),
                 notes: notes_and_auth_paths,
-                recipient_amounts
+                output_amounts
             }
         }
     }
@@ -899,6 +1000,7 @@ mod tests {
 
     use super::Builder;
     use crate::{
+        builder::BundleType,
         bundle::{Authorized, Bundle, Flags},
         circuit::ProvingKey,
         constants::MERKLE_DEPTH_ORCHARD,
@@ -916,19 +1018,20 @@ mod tests {
         let fvk = FullViewingKey::from(&sk);
         let recipient = fvk.address_at(0u32, Scope::External);
 
-        let mut builder = Builder::new(
+        let mut builder = Builder::new(BundleType::Transactional(
             Flags::from_parts(true, true),
             EMPTY_ROOTS[MERKLE_DEPTH_ORCHARD].into(),
-        );
+        ));
 
         builder
-            .add_recipient(None, recipient, NoteValue::from_raw(5000), None)
+            .add_output(None, recipient, NoteValue::from_raw(5000), None)
             .unwrap();
         let balance: i64 = builder.value_balance().unwrap();
         assert_eq!(balance, -5000);
 
         let bundle: Bundle<Authorized, i64> = builder
             .build(&mut rng)
+            .unwrap()
             .unwrap()
             .create_proof(&pk, &mut rng)
             .unwrap()